+++ bibfor/third_party_interf/breselu.F90
@@ -109,7 +109,7 @@ subroutine breselu(typco, alphacc, effmy, effmz, effn, &
 #include "extern/dintelu.h"
 #include "asterc/r8prem.h"
 !
-    integer :: typco
+    integer(kind=8) :: typco
     real(kind=8) :: alphacc
     real(kind=8) :: effmy
     real(kind=8) :: effmz
@@ -124,15 +124,15 @@ subroutine breselu(typco, alphacc, effmy, effmz, effn, &
     real(kind=8) :: fbeton
     real(kind=8) :: gammas
     real(kind=8) :: gammac
-    integer :: clacier
+    integer(kind=8) :: clacier
     real(kind=8) :: eys
-    integer :: typdiag
-    integer :: ferrcomp
-    integer :: precs
-    integer :: ferrsyme
+    integer(kind=8) :: typdiag
+    integer(kind=8) :: ferrcomp
+    integer(kind=8) :: precs
+    integer(kind=8) :: ferrsyme
     real(kind=8) :: slsyme
-    integer :: uc
-    integer :: um
+    integer(kind=8) :: uc
+    integer(kind=8) :: um
     real(kind=8) :: dnsyi
     real(kind=8) :: dnsys
     real(kind=8) :: dnszi
@@ -147,11 +147,11 @@ subroutine breselu(typco, alphacc, effmy, effmz, effn, &
     real(kind=8) :: eczs
     real(kind=8) :: alphay
     real(kind=8) :: alphaz
-    integer :: pivoty
-    integer :: pivotz
-    integer :: etaty
-    integer :: etatz
-    integer :: ierr
+    integer(kind=8) :: pivoty
+    integer(kind=8) :: pivotz
+    integer(kind=8) :: etaty
+    integer(kind=8) :: etatz
+    integer(kind=8) :: ierr
 
 !-----------------------------------------------------------------------
 !!!!VARIABLES DE CALCUL
@@ -160,12 +160,12 @@ subroutine breselu(typco, alphacc, effmy, effmz, effn, &
     real(kind=8) :: rhoyinf, rhoysup, rhozinf, rhozsup
     real(kind=8) :: BRES, mrdyE, mrdy1, mrdy2, mrdzE, mrdz1, mrdz2, nrdyzE, a, nrd0
     logical :: COND
-    integer :: s, COUNT_BRES
+    integer(kind=8) :: s, COUNT_BRES
     real(kind=8), pointer :: nrdy(:) => null(), mrdy(:) => null()
     real(kind=8), pointer :: nrdz(:) => null(), mrdz(:) => null()
     character(24) :: pnrdy, pmrdy, pnrdz, pmrdz
     real(kind=8) :: unite_m, seuil_moment
-    integer :: ntoty, ndemiy, ntotz, ndemiz
+    integer(kind=8) :: ntoty, ndemiy, ntotz, ndemiz
 
     pnrdy = 'POINT_NRD_Y'
     pmrdy = 'POINT_MRD_Y'
